prevent admin from deleting self

+++ flask/app/users.py
@@ -209,6 +209,8 @@ def create_user():
 @check_admin
 def delete_user(username: str):
     user = models.User.query.filter_by(username=username).first_or_404()
+    if user.is_admin:
+        abort(422, description="Admin cannot delete self")
     try:
         db.session.delete(user)
         db.session.commit()
